==> Instacart Basket Analysis Github/03 Scripts/1_Data_Basic_Statistical_Analysis_Wrangling.ipynb <==
# Evaluating Dataframes and Wrangling

Table of Contents 
1. Importation and Basic Statistical Analysis\
   a. Import Libraries\
   b. Import Data\
   c. Check Size, Shape, and Basic Statistical Analysis
2. Data Wrangling\
   a. Drop Columns\
   d. Rename Columns\
   c. Change Datatypes
3. Import Departments Data\
   a. Import Dataset\
   b. Transpose Dataset\
   c. Change Header
4. Departments Data Dictionary\
   a. Create a Dictionary\
   b. Subset Using Data Dictionary
5. Export Dataframes


## 1. Importation and Basic Statistical Analysis

### a. Import Libraries

# import libraries
import pandas as pd
import numpy as np
import os

### b. Import Data

# data folder path
path = r'/Users/catherinemikelson/Desktop/Career Foundry/Acheivement 4 - Python Fundmentals/Instacart Basket Analysis'

# orders dataframe
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

# products dataframe
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

### c. Check Size, Shape, Basic Statistical Analysis

df_ords.head(20)

df_ords.shape

df_ords.info()

df_ords.dtypes

df_ords.describe()

df_prods.head(20)

df_prods.shape

df_prods.info()

df_prods.dtypes

df_prods.describe()

Observations: There may be an error in price as the maximum is 99,999.

## 2. Data Wrangling

### a. Drop Columns

The df_ord column eval_set is not necessary and can be removed.

# drop eval_set from df_ords
df_ords = df_ords.drop(columns = ['eval_set'])

# check output
df_ords.head()

### b. Rename Columns

Change order_dow to orders_day_of_the_week for readability.

# rename order_dow = orders_day_of_the_week
df_ords.rename(columns = {'order_dow' : 'orders_day_of_the_week'}, inplace = True)

### c. Change Data Types

Change df_ords column order_id to a string

# change 'order_id' datatype
df_ords['order_id'] = df_ords['order_id'].astype('str')

# check type
df_ords['order_id'].dtype

## 3. Import Departments Data

### a. Import Dataset

# Import Deptartments File
df_dep = pd.read_csv(os.path.join(path, '02 Data','Original Data', 'departments.csv'), index_col = False)

df_dep.head()

### b. Transpose Dataset

# transpose df_dep
df_dep_t = df_dep.T

df_dep_t.head()

# add an index
df_dep_t.reset_index()

### c. Change Header

The header needs to be reset to the top of the frame.

# take the first row of df_dep_t for the header
new_header = df_dep_t.iloc[0]

# take the data under the header row for a new df
df_dep_t_new = df_dep_t[1:]

# set the header row as the df header
df_dep_t_new.columns = new_header

# check output
df_dep_t_new

## 4. Departments Data Dictionary

### a. Create Data Dictionary

# create a df_dept dictionary
data_dict = df_dep_t_new.to_dict('index')

# check output
data_dict

# check df_prods to associate the dictionary
df_prods.head()

### b. Subset Using Data Dictionary

# check output for dept 19 
print(data_dict.get('19'))

# create df_snacks
df_snacks =  df_prods[df_prods['department_id']==19]

# check output
df_snacks.head()

# create subset using multiple values
df_snacks_3 = df_prods.loc[df_prods['department_id'].isin([17,18,19])]

# check output
df_snacks_3.head()

## 5. Export Dataframes

path_git = r'/Users/catherinemikelson/Desktop/Career Foundry/Acheivement 4 - Python Fundmentals/Instacart Basket Analysis Github'

# Export df_ords as orders_wrangled_git.csv
df_ords.to_csv(os.path.join(path, '02 Data', 'orders_wrangled_git.csv'))

# Export df_dep_t_new as departments_wrangled_git.csv”
df_dep_t_new.to_csv(os.path.join(path, '02 Data', 'departments_wrangled_git.csv'))
