==> Instacart Basket Analysis Github/03 Scripts/2_Data_Consistency.ipynb <==
# Data Consistency and Cleaning

1. Import Libraries and Dataframes
2. Missing Values\
   a. Prods Data\
   b. Ords Data
3. Duplicate Values\
   a. Prods Data\
   b. Ords Data
4. Mixed Data Type
5. Additional Exploratory Analysis
6. Export Data

## 1. Import Libraries and Dataframes

# Import Libraries
import pandas as pd
import numpy as np
import os

# data folder path
path = r'/Users/catherinemikelson/Desktop/Career Foundry/Acheivement 4 - Python Fundmentals/Instacart Basket Analysis'

# Data folder path_git
path_git = r'/Users/catherinemikelson/Desktop/Career Foundry/Acheivement 4 - Python Fundmentals/Instacart Basket Analysis Github'

# Products dataframe
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

# Orders wrangled dataframe
df_ords = pd.read_csv(os.path.join(path_git, '02 Data', 'orders_wrangled_git.csv'))

df_prods.head()

df_ords.head()

# drop unnamed columns
df_ords = df_ords.drop(columns = ['Unnamed: 0'])

# check output 
df_ords.head()

## 2. Missing Values Prods Data

### a. Prods Data

# check for missing values
df_prods.isnull().sum()

# find missing values
df_nan = df_prods[df_prods['product_name'].isnull() == True]

df_nan

Observation: There are 16 missing values. They are few in comparison to the dataframe size and will be removed.

# check shape of frame before removing missing values
df_prods.shape

# remove missing values
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

# check cleaned df shape
df_prods_clean.shape

### b. Ords Data

# look for missing values in df_ords data
df_ords.isnull().sum()

# create dataframe with only null values
df_nan_ords = df_ords[df_ords['days_since_prior_order'].isnull() == True]

df_nan_ords.head(30)

Observation: The null values align with customers who have placed 1 order.  It's logical that new customers would not yet have a value under "days_since_prior_order."  These values can be left as they are.

## 3. Duplicates

### a. Prods Data

# filter duplicates
df_dups_prods = df_prods_clean[df_prods_clean.duplicated()]

df_dups_prods

Observation: There are 5 missing values.  They are few in comparison to the dataframe size and will be removed.

# check the dataframe size before removing duplicates
df_prods_clean.shape

# df with no duplicates
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

# check shape
df_prods_clean_no_dups.shape

### b. Ords Data

# filter duplicates
df_dups_ords = df_ords[df_ords.duplicated()]

# check output
df_dups_ords

Observations: There are no duplicates in the ords data.

### c. Create New Customer

# create new_customer column based on null values for 'days_since_prior_order'
df_ords['new_customer'] = df_ords['days_since_prior_order'].isnull() == True

# check output
df_ords.head()

## 4. Mixed Type Data

for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

## 5. Additional Exploratory Analysis

# df_ords descriptive statistics
df_ords.describe()

Observation: The statistics look as I would expect.  The minimums and maximums make sense with a 7 day week (recorded using values 0-6) and 24 hour day (recorded using values 0-23).  The averages are plausible.

## 6. Export Data

# Export Ords
df_ords.to_csv(os.path.join(path_git, '02 Data', 'orders_cleaned_git.csv'))

# Export Prods
df_prods_clean_no_dups.to_csv(os.path.join(path_git, '02 Data', 'prods_cleaned_git.csv'))
