==> Instacart Basket Analysis Github/03 Scripts/3_Merging_Dataframes.ipynb <==
# Merging Data

1. Import Libraries and Dataframes
2. Merge df_ords and df_ords_prior
3. Merge df_merged_large and df_products¶
4. Export Merged Dataframe

## 1. Import Libraries/Dataframes

# import libraries 
import pandas as pd
import numpy as np
import os

# data folder path
path_git = r'/Users/catherinemikelson/Desktop/Career Foundry/Acheivement 4 - Python Fundmentals/Instacart Basket Analysis Github'

# data folder path
path = r'/Users/catherinemikelson/Desktop/Career Foundry/Acheivement 4 - Python Fundmentals/Instacart Basket Analysis'

# Import order_products
df_ords_prior = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'order_products_prior.csv'), index_col = False)

# check output
df_ords_prior.head()

# check shape
df_ords_prior.shape

# import orders_cleaned
df_ords = pd.read_csv(os.path.join(path_git, '02 Data', 'orders_cleaned_git.csv'), index_col = False)

# check output
df_ords.head()

# drop duplicate index
df_ords = df_ords.drop('Unnamed: 0', axis = 1)

# check output
df_ords.head()

# check shape
df_ords.shape

# import products_cleaned
df_products = pd.read_csv(os.path.join(path_git, '02 Data', 'prods_cleaned_git.csv'))

# check output
df_products.head()

# drop index duplicate
df_products = df_products.drop('Unnamed: 0', axis = 1)

df_products.head()

# check shape
df_products.shape

## 2. Merge df_ords and df_ords_prior

Note: df_ords will be merged with df_ords_prior, a new dataframe, in order to add product_id to the dataframe.

# merge df_ords and df_ords_prior
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

# check output
df_merged_large.head()

# check shape
df_merged_large.shape

# check value counts for '_merge' column
df_merged_large['_merge'].value_counts()

Observation: Success! The new data frame has the same number rows and columns as expected from the merge.

## 3. Merge df_merged_large and df_products

Note: Merging in df_prods will give more specific details on products that were purchased.

# merge df_merged_large and df_products
df_merged_large_prods = df_merged_large.merge(df_products, on = 'product_id')

# check output
df_merged_large_prods.head()

# check merge
df_merged_large_prods['_merge'].value_counts()

Observation: Success! The number of rows is consistent with rows pre-merge.

## 4. Export Merged Dataframe

# export dataframe
df_merged_large_prods.to_pickle(os.path.join(path_git, '02 Data', 'ords_prods_merged.pkl'))
