==> Instacart Basket Analysis Github/03 Scripts/4_Deriving_New_Variables.ipynb <==
# Deriving New Variables

1. Import Libraries and Dataframe
2. Add Price Label
3. Create New Vairables
    a. Create Variable for Busiest/Least Busy Day\
    b. Create Variable for Two Busiest and Two Least Busy Days
    c. Create Variable for Busiest Periods of Day
4. Export Data

## 1. Import Libraries and Dataframe

# import libraries
import pandas as pd
import numpy as np
import os

# data folder path
path_git = r'/Users/catherinemikelson/Desktop/Career Foundry/Acheivement 4 - Python Fundmentals/Instacart Basket Analysis Github'

# import ords/prods
ords_prods_merged = pd.read_pickle(os.path.join(path_git, '02 Data', 'ords_prods_merged.pkl'))

# check output
ords_prods_merged.head()

## 2. Add Price Label

Note: The price label will group products by price range - low, mid or high.

# use locate to create price_range_loc
ords_prods_merged.loc[ords_prods_merged['prices'] > 15, 'price_range_loc'] = 'High-range product'

ords_prods_merged.loc[(ords_prods_merged['prices'] <= 15) & (ords_prods_merged['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 

ords_prods_merged.loc[ords_prods_merged['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

# check value counts
ords_prods_merged['price_range_loc'].value_counts(dropna = False)

Observation: Most products fit into the mid-range, followed by low and high-range respectively.

ords_prods_merged.head()

## 3. Create Busiest Day(s) Variables

### a. Create Variable for Busiest/Least Busy Day

Note: Which day is the busiest?  Least busy?\
Saturday = 0, Sunday = 1, Monday = 2, Tuesday = 3, Wednesday = 4, Thursday = 5, Friday = 6

# check counts of busiest days
ords_prods_merged['orders_day_of_the_week'].value_counts(dropna = False)

# create result and define values
result = []

for value in ords_prods_merged["orders_day_of_the_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

# apply result to ords_prods_merged dataframe
ords_prods_merged['busiest_day'] = result

# check output
ords_prods_merged.head(10)

# check value counts for busiest days
ords_prods_merged['busiest_day'].value_counts(dropna = False)

Observations: Saturday is the busiest day.  Wednesday is the least busy.

### b. Create Variable for Two Busiest and Two Least Busy Days

Note: We are now looking for the TWO busiest and least busy days.  The remainder will be marked as regularly busy.

# create result and define values for Busiest days
result = []

for value in ords_prods_merged["orders_day_of_the_week"]:
  if value == 0 or value == 1:
    result.append("Busiest days")
  elif value == 3 or value == 4:
    result.append("Least busy days")
  else:
    result.append("Regularly busy days")

# assign result
ords_prods_merged['busiest_days'] = result

# check output
ords_prods_merged.head(10)

# check value counts on result
ords_prods_merged['busiest_days'].value_counts(dropna = False)

Note: Saturday and Sunday are the two busiest days.  Tuesday and Wednesday are least busy.  The remainder are regularly busy.

### 3. Create Busiest Periods of Day

Note: The app is dealing with user overload during busiest periods.  Define time periods based on customer use.

# check value counts for order_hour_of_day
ords_prods_merged['order_hour_of_day'].value_counts(dropna = False)

# create result and define values for busiest hour
result = []

for value in ords_prods_merged['order_hour_of_day']:
  if value in [9, 10, 11, 12, 13, 14, 15, 16]:
    result.append('most_orders')
  elif value in [1, 2, 3, 4, 5]:
    result.append('fewest_orders')
  else:
    result.append('average_orders')

# assign result
ords_prods_merged['busiest_hour'] = result

# check for new column
ords_prods_merged.head(10)

#value count
ords_prods_merged['busiest_hour'].value_counts(dropna = False)

Note: The period from 9am-4pm has the most orders.  1am-5am has the least orders.  The remaining time period is marked as an average number of orders.

## 4. Export Data

# Export Data Frame
ords_prods_merged.to_pickle(os.path.join(path_git,'02 Data', 'ords_prods_merged_2.pkl'))
